# resources/openapi/georchestra-plui-evolution.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Georchestra - Plugin Plui'Evolution' API",
        "contact": {
            "email": "thomas.brule@open-groupe.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
	"servers": [
		{
			"url": "/pluievolution",
			"description": "Georchestra - Plugin PluiEvolution API"
		}
	],    
...
    "paths": {
...
        "/request/{uuid}/upload": {
            "post": {
                "tags": [
                    "request"
                ],
                "summary": "Téléversement d'une pièce jointe sur une demande",
                "operationId": "uploadDocument",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Identifiant de la demande",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "description": "Pièce jointe sous la forme d'un blob",
                                        "format": "binary"
                                    }
                                }
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "description": "Pièce jointe sous la forme d'un blob",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/Attachment"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "bad input parameter",
                        "content": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/request/{uuid}/attachments": {
            "get": {
                "tags": [
                    "request"
                ],
                "summary": "Récupération de la liste des pièces jointes d'une demande",
                "operationId": "getAttachementsByPluiRequestUuid",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Identifiant de la demande",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Attachment"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "bad input parameter",
                        "content": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/request": {
            "put": {
                "tags": [
                    "request"
                ],
                "summary": "Maj d'une demande",
                "description": "Permet de mettre à jour une demande",
                "operationId": "updatePluiRequest",
                "requestBody": {
                    "description": "Demande à metre à jour",
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/PluiRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Demande mise à jour",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/PluiRequest"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "x-codegen-request-body-name": "pluiRequest"
            },
            "post": {
                "tags": [
                    "request"
                ],
                "summary": "Création d'une nouvelle demande",
                "description": "Permet de créer une nouvelle demande",
                "operationId": "createPluiRequest",
                "requestBody": {
                    "description": "Demande à créer",
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/PluiRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Demande créée",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/PluiRequest"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "x-codegen-request-body-name": "pluiRequest"
            }
        },
        "/request/{uuid}": {
            "get": {
                "tags": [
                    "request"
                ],
                "summary": "Récupération d'une demande spécifique en fonction de son uuid",
                "description": "Permet d'obtenir une demande à partir de son uuid",
                "operationId": "getPluiRequestByUuid",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Uuid de la demande",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Demande trouvée",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/PluiRequest"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "request"
                ],
                "summary": "Suppression d'une demande",
                "description": "Permet de supprimer une demande à partir de son uuid",
                "operationId": "deletePluiRequest",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Uuid de la demande à supprimer",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {}
                    }
                }
            }
        },
...
        "/geographic/etablissements": {
            "get": {
                "tags": [
                    "geographic"
                ],
                "summary": "Liste des mairies de communes, plus ville et métropole, d'où les demandes peuvent être émises",
                "operationId": "searchEtablissements",
                "responses": {
                    "200": {
                        "description": "Liste des mairies de communes + ville + metropole",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GeographicEtablissement"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/geographic/configuration": {
            "get": {
                "tags": [
                    "geographic"
                ],
                "summary": "Configuration des établissements",
                "operationId": "getEtablissementConfiguration",
                "responses": {
                    "200": {
                        "description": "Configuration des établissements",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/EtablissementConfiguration"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/geographic/areas": {
            "get": {
                "tags": [
                    "geographic"
                ],
                "summary": "Liste des emprises de communes, plus ville et métropole, d'où les demandes peuvent être émises",
                "operationId": "getAllPluiRequestAreas",
                "responses": {
                    "200": {
                        "description": "Liste des emprises de communes + ville + metropole",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GeographicArea"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/geographic/areas/{codeInsee}": {
            "get": {
                "tags": [
                    "geographic"
                ],
                "summary": "Emprises d'une communes ou ville où métropole à partir de son code Insee",
                "operationId": "getPluiRequestAreaByCodeInsee",
                "parameters": [
                    {
                        "name": "codeInsee",
                        "in": "path",
                        "description": "Code insee de la geographique area",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Emprises de commune | ville | metropole",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeographicArea"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/geographic/etablissements/{codeInsee}": {
            "get": {
                "tags": [
                    "geographic"
                ],
                "summary": "Hotel de ville d'une commune (ou Rennes Metropole) par son code insee",
                "operationId": "getPluiRequestEtablissementByCodeInsee",
                "parameters": [
                    {
                        "name": "codeInsee",
                        "in": "path",
                        "description": "Code insee du geographique etablissement",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hotel de ville | metropole",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeographicEtablissement"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ApiError": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "label": {
                        "type": "string"
                    }
                }
            },
...
            "Attachment": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "mimeType": {
                        "type": "string"
                    }
                }
            },
...
            "EtablissementConfiguration": {
                "type": "object",
                "properties": {
                    "organisationRm": {
                        "type": "string"
                    },
                    "codeInseeRm": {
                        "type": "string"
                    }
                }
            },
...
            "GeometryType": {
                "type": "string",
                "description": "the geometry type",
                "enum": [
                    "Point",
                    "LineString",
                    "Polygon",
                    "MultiPoint",
                    "MultiLineString",
                    "MultiPolygon"
                ]
            },
            "Geometry": {
                "type": "object",
                "properties": {
                    "type": {
                        "$ref": "#/components/schemas/GeometryType"
                    },
                    "projection": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "description": "GeoJSon geometry",
                "discriminator": {
                    "propertyName": "type"
                }
            },
            "Point2D": {
                "maxItems": 2,
                "minItems": 2,
                "type": "array",
                "items": {
                    "type": "number"
                }
            },
            "Point": {
                "description": "GeoJSon geometry",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Geometry"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "coordinates": {
                                "$ref": "#/components/schemas/Point2D"
                            }
                        }
                    }
                ]
            },
...
            "Polygon": {
                "description": "GeoJSon geometry",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Geometry"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "coordinates": {
                                "type": "array",
                                "items": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Point2D"
                                    }
                                }
                            }
                        }
                    }
                ]
            },
...
            "PluiRequest": {
                "type": "object",
                "properties": {
                    "uuid": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "subject": {
                        "type": "string"
                    },
                    "object": {
                        "type": "string"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "pluiProcedure": {
                        "type": "string"
                    },
                    "concertation": {
                        "type": "string"
                    },
                    "approbation": {
                        "type": "string"
                    },
                    "initiator": {
                        "type": "string"
                    },
                    "creationDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "localisation": {
                        "$ref": "#/components/schemas/Point"
                    },
                    "status": {
                        "$ref": "#/components/schemas/PluiRequestStatus"
                    },
                    "type": {
                        "$ref": "#/components/schemas/PluiRequestType"
                    },
                    "codeInsee": {
                        "type": "string"
                    },
                    "redmineId": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "description": "Il s'agit de la representation d'une demande"
            },
            "Geographic": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nom": {
                        "type": "string"
                    },
                    "codeInsee": {
                        "type": "string"
                    }
                }
            },
            "GeographicEtablissement": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Geographic"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "localisation": {
                                "$ref": "#/components/schemas/Point"
                            }
                        }
                    }
                ]
            },
            "GeographicArea": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Geographic"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "localisation": {
                                "$ref": "#/components/schemas/Polygon"
                            },
                            "identifiantRedmine": {
                                "type": "string"
                            }
                        }
                    }
                ]
            },
            "PluiRequestStatus": {
                "type": "string",
                "enum": [
                    "NOUVEAU",
                    "ANALYSE_EN_COURS",
                    "PRODUCTION_EN_COURS",
                    "EN_ATTENTE_VALIDATION_COMMUNE",
                    "VALIDE_COMMUNE",
                    "DEMANDE_NON_RECEVABLE",
                    "DEMANDE_REFORMULEE"
                ]
            },
            "PluiRequestType": {
                "type": "string",
                "enum": [
                    "COMMUNE",
                    "INTERCOMMUNE",
                    "METROPOLITAIN"
                ]
            }
        }
    }
}